Fix uint8(int256) to uint8(uint256(int256))

+++ packages/resolver/solidity/AssertIntArray.sol
@@ -173,7 +173,7 @@ library AssertIntArray {
             neg = true;
         }
         while (n > 0) {
-            bts[i++] = _utoa(uint8(n % radix)); // Turn it to ascii.
+            bts[i++] = _utoa(uint8(uint(n % radix))); // Turn it to ascii.
             n /= radix;
         }
         // Reverse
